requirements.txt
alembic==1.8.1
anyio==3.6.2
appnope==0.1.3
asttokens==2.1.0
backcall==0.2.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
decorator==5.1.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
executing==1.2.0
fastapi==0.87.0
h11==0.14.0
httpcore==0.16.1
httptools==0.5.0
httpx==0.23.1
idna==3.4
ipdb==0.13.9
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
orjson==3.8.1
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
stack-data==0.6.1
starlette==0.21.0
toml==0.10.2
traitlets==5.5.0
typing_extensions==4.4.0
ujson==5.5.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
